Add root lock check for Arch Linux
The 'arch-install' installer includes an option to lock out the root user and use a sudo only system. When that's the case the PTS fails to install Arch Linux packages. This uses sudo passwd -S to check if the root account is locked and runs pacman with sudo instead of su -c if true.