Added support for HTTPS Phoromatic Servers
This addresses issue #558.  It is a little rudimentary, but if the port is 443 it will leverage the HTTPS protocol.  In all other cases it will leverage HTTP.

Revert "Added support for HTTPS Phoromatic Servers"

This reverts commit 4b24fec76062b398855f1446624ad4f3ce7ef994.

Added support for HTTPS Phoromatic Servers

This addresses issue #558. It is a little rudimentary, but if the port is 443 it will leverage the HTTPS protocol. In all other cases it will leverage HTTP.

Update phoromatic.php

Changes from @michaellarabel

Revert "Update phoromatic.php"

This reverts commit 7a4a611dc6e652f2323e845582b5393098c9c92f.

Applied Fixes

Applied fixes from @michaellarabel with update to line 370